chore: :truck: move `constants.js` file at root of /implementations
